--- FrontEnd/src/components/Login/LoginBox.tsx
@@ -1,11 +1,0 @@
-import * as styles from "@/components/Login/LoginBox.style";
-
-const LoginBox = () => {
-  return (
-    <styles.Wrapper>
-      <h5>로그인 페이지</h5>
-    </styles.Wrapper>
-  );
-};
-
-export default LoginBox;
