+++ FrontEnd/src/components/main/Login/LoginBox.style.tsx
@@ -1,0 +1,11 @@
+import styled, { css } from "styled-components";
+
+export const Wrapper = styled.div`
+  ${({ theme }) => {
+    const { color, font } = theme;
+    return css`
+      display: flex;
+      flex-direction: column;
+    `;
+  }}
+`;
