+++ FrontEnd/src/components/main/Login/LoginBox.tsx
@@ -1,0 +1,11 @@
+import * as styles from "~/src/components/main/Login/LoginBox.style";
+
+const LoginBox = () => {
+  return (
+    <styles.Wrapper>
+      <h5>로그인 페이지</h5>
+    </styles.Wrapper>
+  );
+};
+
+export default LoginBox;
